+++ mobile/android/android-components/components/browser/engine-system/src/test/java/mozilla/components/browser/engine/system/matcher/UrlMatcherTest.kt
@@ -184,22 +184,6 @@ class UrlMatcherTest {
     }
     """
 
-    val OVERRIDES = """{
-      "categories": {
-        "Advertising": [
-          {
-            "AdTest2": {
-              "http://www.adtest2.com/": [
-                "adtest2.de",
-                "adtest2.at"
-              ]
-            }
-          }
-        ]
-      }
-    }
-    """
-
     val WHITE_LIST = """{
       "SocialTest1": {
         "properties": [
@@ -214,7 +198,6 @@ class UrlMatcherTest {
     fun createMatcher() {
         val matcher = UrlMatcher.createMatcher(
                 StringReader(BLOCK_LIST),
-                listOf(StringReader(OVERRIDES)),
                 StringReader(WHITE_LIST))
 
         // Check returns correct category
@@ -252,11 +235,6 @@ class UrlMatcherTest {
         assertTrue(matchesAnalytics)
         assertEquals(categoryAnalytics, ANALYTICS)
 
-        // Check that override worked
-        assertTrue(matcher.matches("http://adtest2.com", "http://www.adtest2.com").first)
-        assertTrue(matcher.matches("http://adtest2.at", "http://www.adtest2.com").first)
-        assertTrue(matcher.matches("http://adtest2.de", "http://www.adtest2.com").first)
-
         // Check that white list worked
         assertTrue(matcher.matches("http://socialtest1.com", "http://www.socialtest1.com").first)
         assertFalse(matcher.matches("http://socialtest1.de", "http://www.socialtest1.com").first)
@@ -281,7 +259,6 @@ class UrlMatcherTest {
     fun setCategoriesEnabled() {
         val matcher = spy(UrlMatcher.createMatcher(
                 StringReader(BLOCK_LIST),
-                listOf(StringReader(OVERRIDES)),
                 StringReader(WHITE_LIST),
                 setOf("Advertising", "Analytics"))
         )
@@ -299,7 +276,6 @@ class UrlMatcherTest {
     fun webFontsNotBlockedByDefault() {
         val matcher = UrlMatcher.createMatcher(
                 StringReader(BLOCK_LIST),
-                listOf(StringReader(OVERRIDES)),
                 StringReader(WHITE_LIST),
                 setOf(UrlMatcher.ADVERTISING, UrlMatcher.ANALYTICS, UrlMatcher.SOCIAL, UrlMatcher.CONTENT))